fix comments, separate multiple classes in single file to different files, make class structure same
