+++ src/Group/Group.php
@@ -4,26 +4,6 @@ namespace JiraRestApi\Group;
 
 use JiraRestApi\ClassSerialize;
 
-class GroupUser
-{
-    /**
-     * @var int
-     */
-    public $size;
-
-    /** @var array */
-    public $items;
-
-    /** @var int */
-    public $max_results;
-
-    /** @var int */
-    public $start_index;
-
-    /** @var int */
-    public $end_index;
-}
-
 /**
  * Class Group.
  *
