+++ src/Group/GroupUser.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace JiraRestApi\Group;
+
+class GroupUser
+{
+    /**
+     * @var int
+     */
+    public $size;
+
+    /** @var array */
+    public $items;
+
+    /** @var int */
+    public $max_results;
+
+    /** @var int */
+    public $start_index;
+
+    /** @var int */
+    public $end_index;
+}
